Optimize sandbox lifecycle management
1. Add many roll back handlers in the process of starting sandbox.
2. Move setup network to Start sandbox as it should be called near the vm startup.
3. Only monit running sandbox and dump it status in time.
4. Sandbox stop should make sure the sandbox is stopped successfully, so wait it
to stop for 10s.
5. If the VM process terminated suddenly and the containerd has no idea to stop
sandbox, destroy network should be done in monit thread.


Match-id-d6574b74eae6412f6879e55d58e952935fd6b336

+++ vmm/sandbox/src/cloud_hypervisor/mod.rs
@@ -19,7 +19,7 @@ use std::{os::unix::io::RawFd, process::Stdio};
 use anyhow::anyhow;
 use async_trait::async_trait;
 use containerd_sandbox::error::{Error, Result};
-use log::{debug, error, info};
+use log::{debug, error, info, warn};
 use nix::{errno::Errno::ESRCH, sys::signal, unistd::Pid};
 use serde::{Deserialize, Serialize};
 use time::OffsetDateTime;
@@ -45,7 +45,7 @@ use crate::{
     impl_recoverable, load_config,
     param::ToCmdLineParams,
     sandbox::KuasarSandboxer,
-    utils::{read_std, set_cmd_fd, set_cmd_netns, wait_pid, write_file_atomic},
+    utils::{read_std, set_cmd_fd, set_cmd_netns, wait_channel, wait_pid, write_file_atomic},
     vm::{Pids, VcpuThreads, VM},
 };
 
@@ -146,6 +146,17 @@ impl CloudHypervisorVM {
         self.fds.push(fd);
         self.fds.len() - 1 + 3
     }
+
+
+    async fn wait_stop(&mut self, t: Duration) -> Result<()> {
+        if let Some(rx) = self.wait_channel().await {
+            let (_, ts) = *rx.borrow();
+            if ts == 0 {
+                wait_channel(t, rx).await?;
+            }
+        }
+        Ok(())
+    }
 }
 
 #[async_trait]
@@ -153,6 +164,8 @@ impl VM for CloudHypervisorVM {
     async fn start(&mut self) -> Result<u32> {
         create_dir_all(&self.base_dir).await?;
         let virtiofsd_pid = self.start_virtiofsd().await?;
+        // TODO: add child virtiofsd process
+        self.pids.affiliated_pids.push(virtiofsd_pid);
         let mut params = self.config.to_cmdline_params("--");
         for d in self.devices.iter() {
             params.extend(d.to_cmdline_params("--"));
@@ -175,21 +188,27 @@ impl VM for CloudHypervisorVM {
             .spawn()
             .map_err(|e| anyhow!("failed to spawn cloud hypervisor command: {}", e))?;
         let pid = child.id();
+        self.pids.vmm_pid = pid;
         let pid_file = format!("{}/pid", self.base_dir);
-        let (tx, rx) = tokio::sync::watch::channel((0u32, 0i128));
+        let (tx, rx) = channel((0u32, 0i128));
+        self.wait_chan = Some(rx);
         spawn_wait(
             child,
             format!("cloud-hypervisor {}", self.id),
             Some(pid_file),
             Some(tx),
         );
-        self.client = Some(self.create_client().await?);
-        self.wait_chan = Some(rx);
 
-        // update vmm related pids
-        self.pids.vmm_pid = pid;
-        self.pids.affiliated_pids.push(virtiofsd_pid);
-        // TODO: add child virtiofsd process
+        match self.create_client().await {
+            Ok(client) => self.client = Some(client),
+            Err(e) => {
+                if let Err(re) = self.stop(true).await {
+                    warn!("roll back in create clh api client: {}", re);
+                    return Err(e);
+                }
+                return Err(e);
+            }
+        };
         Ok(pid.unwrap_or_default())
     }
 
@@ -205,10 +224,12 @@ impl VM for CloudHypervisorVM {
             if vmm_pid > 0 {
                 // TODO: Consider pid reused
                 match signal::kill(Pid::from_raw(vmm_pid as i32), signal) {
-                    Err(e) if e != ESRCH => {
-                        return Err(anyhow!("kill vmm process {}: {}", vmm_pid, e).into());
+                    Err(e) => {
+                        if e != ESRCH {
+                            return Err(anyhow!("kill vmm process {}: {}", vmm_pid, e).into());
+                        }
                     }
-                    _ => {}
+                    Ok(_) => self.wait_stop(Duration::from_secs(10)).await?,
                 }
             }
         }
